Move back to a single deprecations file and reorder in array
