--- packages/block-library/src/gallery/deprecated/v2/index.js
@@ -1,167 +1,0 @@
-/**
- * External dependencies
- */
-import { map } from 'lodash';
-
-/**
- * Internal dependencies
- */
-import { defaultColumnsNumber } from '../../shared';
-
-/**
- * WordPress dependencies
- */
-import { RichText } from '@wordpress/block-editor';
-
-export default {
-	attributes: {
-		images: {
-			type: 'array',
-			default: [],
-			source: 'query',
-			selector: '.blocks-gallery-item',
-			query: {
-				url: {
-					source: 'attribute',
-					selector: 'img',
-					attribute: 'src',
-				},
-				fullUrl: {
-					source: 'attribute',
-					selector: 'img',
-					attribute: 'data-full-url',
-				},
-				link: {
-					source: 'attribute',
-					selector: 'img',
-					attribute: 'data-link',
-				},
-				alt: {
-					source: 'attribute',
-					selector: 'img',
-					attribute: 'alt',
-					default: '',
-				},
-				id: {
-					source: 'attribute',
-					selector: 'img',
-					attribute: 'data-id',
-				},
-				caption: {
-					type: 'string',
-					source: 'html',
-					selector: '.blocks-gallery-item__caption',
-				},
-			},
-		},
-		ids: {
-			type: 'array',
-			default: [],
-		},
-		columns: {
-			type: 'number',
-		},
-		caption: {
-			type: 'string',
-			source: 'html',
-			selector: '.blocks-gallery-caption',
-		},
-		imageCrop: {
-			type: 'boolean',
-			default: true,
-		},
-		linkTo: {
-			type: 'string',
-			default: 'none',
-		},
-	},
-	supports: {
-		align: true,
-	},
-	isEligible( { ids } ) {
-		return ids && ids.some( ( id ) => typeof id === 'string' );
-	},
-	migrate( attributes ) {
-		return {
-			...attributes,
-			ids: map( attributes.ids, ( id ) => {
-				const parsedId = parseInt( id, 10 );
-				return Number.isInteger( parsedId ) ? parsedId : null;
-			} ),
-		};
-	},
-	save( { attributes } ) {
-		const {
-			images,
-			columns = defaultColumnsNumber( attributes?.images.length ),
-			imageCrop,
-			caption,
-			linkTo,
-		} = attributes;
-
-		return (
-			<figure
-				className={ `columns-${ columns } ${
-					imageCrop ? 'is-cropped' : ''
-				}` }
-			>
-				<ul className="blocks-gallery-grid">
-					{ images.map( ( image ) => {
-						let href;
-
-						switch ( linkTo ) {
-							case 'media':
-								href = image.fullUrl || image.url;
-								break;
-							case 'attachment':
-								href = image.link;
-								break;
-						}
-
-						const img = (
-							<img
-								src={ image.url }
-								alt={ image.alt }
-								data-id={ image.id }
-								data-full-url={ image.fullUrl }
-								data-link={ image.link }
-								className={
-									image.id ? `wp-image-${ image.id }` : null
-								}
-							/>
-						);
-
-						return (
-							<li
-								key={ image.id || image.url }
-								className="blocks-gallery-item"
-							>
-								<figure>
-									{ href ? (
-										<a href={ href }>{ img }</a>
-									) : (
-										img
-									) }
-									{ ! RichText.isEmpty( image.caption ) && (
-										<RichText.Content
-											tagName="figcaption"
-											className="blocks-gallery-item__caption"
-											value={ image.caption }
-										/>
-									) }
-								</figure>
-							</li>
-						);
-					} ) }
-				</ul>
-				{ ! RichText.isEmpty( caption ) && (
-					<RichText.Content
-						tagName="figcaption"
-						className="blocks-gallery-caption"
-						value={ caption }
-					/>
-				) }
-			</figure>
-		);
-	},
-};
