--- packages/block-library/src/gallery/deprecated/v5/index.js
@@ -1,99 +1,0 @@
-/**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
- * Internal dependencies
- */
-import { defaultColumnsNumber } from '../../shared';
-
-export default {
-	attributes: {
-		images: {
-			type: 'array',
-			default: [],
-			source: 'query',
-			selector: 'div.wp-block-gallery figure.blocks-gallery-image img',
-			query: {
-				url: {
-					source: 'attribute',
-					attribute: 'src',
-				},
-				alt: {
-					source: 'attribute',
-					attribute: 'alt',
-					default: '',
-				},
-				id: {
-					source: 'attribute',
-					attribute: 'data-id',
-				},
-			},
-		},
-		columns: {
-			type: 'number',
-		},
-		imageCrop: {
-			type: 'boolean',
-			default: true,
-		},
-		linkTo: {
-			type: 'string',
-			default: 'none',
-		},
-		align: {
-			type: 'string',
-			default: 'none',
-		},
-	},
-	supports: {
-		align: true,
-	},
-	save( { attributes } ) {
-		const {
-			images,
-			columns = defaultColumnsNumber( attributes?.images.length ),
-			align,
-			imageCrop,
-			linkTo,
-		} = attributes;
-		const className = classnames( `columns-${ columns }`, {
-			alignnone: align === 'none',
-			'is-cropped': imageCrop,
-		} );
-		return (
-			<div className={ className }>
-				{ images.map( ( image ) => {
-					let href;
-
-					switch ( linkTo ) {
-						case 'media':
-							href = image.url;
-							break;
-						case 'attachment':
-							href = image.link;
-							break;
-					}
-
-					const img = (
-						<img
-							src={ image.url }
-							alt={ image.alt }
-							data-id={ image.id }
-						/>
-					);
-
-					return (
-						<figure
-							key={ image.id || image.url }
-							className="blocks-gallery-image"
-						>
-							{ href ? <a href={ href }>{ img }</a> : img }
-						</figure>
-					);
-				} ) }
-			</div>
-		);
-	},
-};
